Refactor Salsa20BaseIntrinsics
